requirements.txt
fastapi==0.128.0
httpx==0.28.1
uvicorn== 0.40.0
python-dotenv==1.2.1
pytest==9.0.2
pydantic>=2.5,<3.0
pydantic-settings==2.12.0
fastapi-cache2>=0.2.1
redis>=5.0
matplotlib==3.10.8
numpy==2.4.1
